+++ nprogress.js
@@ -168,8 +168,6 @@ window.NProgress = (function() {
   NProgress.render = function(fromStart) {
     if (NProgress.isRendered()) return document.getElementById('nprogress');
 
-    addClass(document.documentElement, 'nprogress-busy');
-
     var progress = document.createElement('div');
     progress.id = 'nprogress';
     progress.innerHTML = '<div class="bar"><div class="peg"></div></div>';
@@ -191,7 +189,6 @@ window.NProgress = (function() {
    */
 
   NProgress.remove = function() {
-    removeClass(document.documentElement, 'nprogress-busy');
     var progress = document.getElementById('nprogress');
     progress && removeElement(progress);
   };
@@ -349,56 +346,6 @@ window.NProgress = (function() {
       }
     }
   })();
-
-  /**
-   * (Internal) Determines if an element or space separated list of class names contains a class name.
-   */
-
-  function hasClass(element, name) {
-    var list = typeof element == 'string' ? element : classList(element);
-    return list.indexOf(' ' + name + ' ') >= 0;
-  }
-
-  /**
-   * (Internal) Adds a class to an element.
-   */
-
-  function addClass(element, name) {
-    var oldList = classList(element),
-        newList = oldList + name;
-
-    if (hasClass(oldList, name)) return;
-
-    // Trim the opening space.
-    element.className = newList.substring(1);
-  }
-
-  /**
-   * (Internal) Removes a class from an element.
-   */
-
-  function removeClass(element, name) {
-    var oldList = classList(element),
-        newList;
-
-    if (!hasClass(element, name)) return;
-
-    // Replace the class name.
-    newList = oldList.replace(' ' + name + ' ', ' ');
-
-    // Trim the opening and closing spaces.
-    element.className = newList.substring(1, newList.length - 1);
-  }
-
-  /**
-   * (Internal) Gets a space separated list of the class names on the element.
-   * The list is wrapped with a single space on each end to facilitate finding
-   * matches within the list.
-   */
-
-  function classList(element) {
-    return (' ' + (element && element.className || '') + ' ').replace(/\s+/gi, ' ');
-  }
 
   /**
    * (Internal) Removes an element from the DOM.